refactor: Update Profile page as per new flow designs
- Add profile header component
- Add border styles for badge and activityItem
- Add route for profile screen
