+++ app/profile/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "expo-router";
+
+export default function () {
+	return (
+		<Stack
+			screenOptions={{
+				headerShown: false,
+			}}
+		/>
+	);
+}
