+++ lib/components/profile/index.ts
@@ -1,0 +1,6 @@
+import { Activity, ActivitySkeleton } from "./activity";
+import Badges from "./Badges";
+import PointsView from "./PointsView";
+import ProfileHeader from "./ProfileHeader";
+
+export { Activity, ActivitySkeleton, Badges, PointsView, ProfileHeader };
